03_PROJECTS/vibe_match/src/services/ai/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9
open-clip-torch==2.26.1
torch==2.4.0
torchvision==0.19.0
Pillow==10.4.0
insightface==0.7.3
onnxruntime==1.19.0
numpy==1.26.4
pydantic==2.9.0
